check for casting error and add test in Terminate

+++ venona/pkg/runtime/runtime.go
@@ -16,6 +16,7 @@ package runtime
 
 import (
 	"encoding/json"
+	"fmt"
 
 	"github.com/codefresh-io/go/venona/pkg/kubernetes"
 	"github.com/codefresh-io/go/venona/pkg/task"
@@ -59,7 +60,11 @@ func (r runtime) TerminateWorkflow(tasks []task.Task) error {
 	for _, task := range tasks {
 		opt := kubernetes.DeleteOptions{}
 		opt.Kind = task.Type
-		b := []byte(task.Spec.(string))
+		s, ok := task.Spec.(string)
+		if !ok {
+			return fmt.Errorf("failed to cast to string")
+		}
+		b := []byte(s)
 		err := json.Unmarshal(b, &opt)
 		err = r.client.DeleteResource(opt)
 		if err != nil {
